+++ odoo_health_check/tests/__init__.py
@@ -5,3 +5,4 @@ from . import test_config
 from . import test_disk_check
 from . import test_disk_alert
 from . import test_pg_report
+from . import test_dashboard